requirements.txt
# UI
streamlit==1.39.0

# Database / ORM / Migration
SQLAlchemy==2.0.35
psycopg[binary,pool]==3.2.3
alembic==1.13.3
psycopg2-binary

# LLM (Gemini)
google-generativeai==0.8.3
langchain==0.3.7
langchain-community==0.3.7
langchain-google-genai==2.0.5

# Data processing / ML
pandas==2.2.3
numpy==1.26.4

# Utils / Config
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
loguru==0.7.2
plotly

# Report export (시스템 패키지 필요 있음)
weasyprint==62.3
